+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added new function `New-DatabricksPythonActivity`.
+
 ## [0.3.1] - 2024-11-05
 
 ### Fixed